Refactor some components
