+++ src/components/MAPI/clusters/Cluster.tsx
@@ -4,8 +4,8 @@ import { Redirect, Switch } from 'react-router';
 import Route from 'Route';
 
 import ClusterDetail from './ClusterDetail';
-import CreateCluster from './CreateCluster';
-import CreateClusterAppForm from './CreateCluster/CreateClusterAppForm';
+import CreateCluster from './CreateCluster/CreateCluster';
+import CreateClusterAppBundles from './CreateCluster/CreateClusterAppBundles';
 import GettingStarted from './GettingStarted';
 
 const Cluster: React.FC<React.PropsWithChildren<{}>> = () => {
@@ -17,7 +17,7 @@ const Cluster: React.FC<React.PropsWithChildren<{}>> = () => {
         path={OrganizationsRoutes.Clusters.New}
       />
       <Route
-        component={CreateClusterAppForm}
+        component={CreateClusterAppBundles}
         exact
         path={OrganizationsRoutes.Clusters.NewViaAppBundles}
       />
